Hopefully a fix of the weird files.copy bug on jenkins

+++ src/test/java/dk/statsbiblioteket/newspaper/promptdomsingester/AbstractSimpleFedoraIngesterTest.java
@@ -8,7 +8,6 @@ import org.testng.annotations.Test;
 
 import javax.xml.bind.JAXBException;
 import java.io.File;
-import java.io.IOException;
 import java.net.MalformedURLException;
 import java.util.regex.Pattern;
 
@@ -19,12 +18,17 @@ import static org.testng.Assert.assertTrue;
  */
 public abstract class AbstractSimpleFedoraIngesterTest {
 
-    protected abstract EnhancedFedora getEnhancedFedora() throws MalformedURLException, JAXBException, PIDGeneratorException;
     String pid;
+
+    protected abstract EnhancedFedora getEnhancedFedora() throws
+                                                          MalformedURLException,
+                                                          JAXBException,
+                                                          PIDGeneratorException;
 
     /**
      * Test that we can ingest a sample batch. Subclasses of this class should appropriate asserts to test that this
      * test does what it should.
+     *
      * @throws Exception
      */
     @Test
@@ -34,20 +38,21 @@ public abstract class AbstractSimpleFedoraIngesterTest {
         File rootTestdataDir = new File(System.getProperty("integration.test.newspaper.testdata"));
         File testSource = new File(rootTestdataDir, "small-test-batch/B400022028241-RT1/");
         File testRoot = new File("/tmp/", "B400022028242-RT1");
-        try {
+
+        if (testRoot.isDirectory()) {
             Files.delete(testRoot);
-        } catch (IOException e){
-            //ignore
         }
-        Files.copy(testSource,testRoot,true);
+
+        Files.copy(testSource, testRoot, true);
 
         assertTrue(testRoot.exists(), testRoot.getAbsolutePath() + " does not exist.");
-        TransformingIteratorForFileSystems iterator =
-                new TransformingIteratorForFileSystems(testRoot, Pattern.quote("."), ".*\\.jp2$", ".md5");
+        TransformingIteratorForFileSystems iterator = new TransformingIteratorForFileSystems(
+                testRoot, Pattern.quote("."), ".*\\.jp2$", ".md5");
 
         String rootPid = ingester.ingest(iterator);
         pid = rootPid;
-        System.out.println("Created object tree rooted at " + rootPid);
+        System.out
+              .println("Created object tree rooted at " + rootPid);
         Files.delete(testRoot);
     }
 }
